feat: add aria-invalid and aria-describedby to new form elements
