+++ packages/theme/src/components/forms/index.ts
@@ -1,4 +1,2 @@
-export * from './form-field';
-export * from './form-input-textarea';
-export * from './form-checkbox-radio';
-export * from './form-select';
+export * from './forms';
+export * from './forms-legacy';
